+++ test/controllers/names_controller_index_test.rb
@@ -55,8 +55,8 @@ class NamesControllerIndexTest < FunctionalTestCase
     query = Query.lookup_and_save(:Name, search_name: search_string)
 
     login
-    params = { q: @controller.get_query_param(query), advanced_search: true }
-    get(:index, params:)
+    get(:index,
+        params: @controller.query_params(query).merge(advanced_search: true))
 
     assert_response(:success)
     assert_select(
@@ -76,8 +76,8 @@ class NamesControllerIndexTest < FunctionalTestCase
            "Test needs a string that has exactly one hit")
 
     login
-    params = { q: @controller.get_query_param(query), advanced_search: true }
-    get(:index, params:)
+    get(:index,
+        params: @controller.query_params(query).merge(advanced_search: true))
     assert_match(name_path(names(:stereum_hirsutum)), redirect_to_url,
                  "Wrong page")
   end
@@ -94,31 +94,30 @@ class NamesControllerIndexTest < FunctionalTestCase
   def test_index_advanced_search_no_hits
     query = oklahoma_query
     login
-    params = { q: @controller.get_query_param(query), advanced_search: true }
-    get(:index, params:)
+    get(:index,
+        params: @controller.query_params(query).merge({ advanced_search: "1" }))
 
     assert_page_title(:NAMES.l)
     assert_flash_text(:runtime_no_matches.l(type: :names.l))
   end
 
-  # This test no longer makes sense with permalinks
-  # def test_index_advanced_search_with_deleted_query
-  #   query = oklahoma_query
-  #   params = { q: @controller.get_query_param(query), advanced_search: true }
-  #   query.record.delete
+  def test_index_advanced_search_with_deleted_query
+    query = oklahoma_query
+    params = @controller.query_params(query).merge(advanced_search: true)
+    query.record.delete
 
-  #   login
-  #   get(:index, params:)
+    login
+    get(:index, params: params)
 
-  #   assert_redirected_to(search_advanced_path)
-  # end
+    assert_redirected_to(search_advanced_path)
+  end
 
   def test_index_advanced_search_error
     query_no_conditions = Query.lookup_and_save(:Name)
 
     login
-    params = { q: @controller.get_query_param(query_no_conditions),
-               advanced_search: true }
+    params = @controller.query_params(query_no_conditions).
+             merge({ advanced_search: true })
     get(:index, params:)
 
     assert_flash_error(:runtime_no_conditions.l)
@@ -358,61 +357,72 @@ class NamesControllerIndexTest < FunctionalTestCase
     end
   end
 
-  def pagination_query_param
+  def pagination_query_params
     query = Query.lookup_and_save(:Name, order_by: :name)
-    @controller.get_query_param(query)
+    @controller.query_params(query)
   end
 
   # None of our standard tests ever actually renders pagination_numbers
   # or letter_pagination_nav.  This tests all the above.
   def test_pagination_page1
     # Straightforward index of all names, showing first 10.
+    query_params = pagination_query_params
     login
-    get(:test_index, params: { num_per_page: 10, q: pagination_query_param })
+    get(:test_index, params: { num_per_page: 10 }.merge(query_params))
+    # print @response.body
     assert_template("names/index")
-
     name_links = css_select(".list-group.name-index a")
     assert_equal(10, name_links.length)
     expected = Name.order(:sort_name, :author).limit(10).to_a
     assert_equal(expected.map(&:id), ids_from_links(name_links))
-
+    # assert_equal(@controller.url_with_query(action: "show",
+    #  id: expected.first.id, only_path: true), name_links.first.url)
     url = @controller.url_with_query(controller: "/names", action: :show,
                                      id: expected.first.id, only_path: true)
-    assert_equal(name_links.first[:href], url)
-
+    assert_not_nil(name_links.first.to_s.index(url))
+    # assert_select("a", text: "1", count: 0)
     assert_link_in_html("Next", controller: "/names",
                                 action: :test_index, num_per_page: 10,
-                                q: pagination_query_param, page: 2)
+                                params: query_params, page: 2)
+    # assert_select("a", text: "Z", count: 0)
+    # assert_link_in_html("A", controller: "/names",
+    #                          action: :test_index, num_per_page: 10,
+    #                          params: query_params, letter: "A")
   end
 
   def test_pagination_page2
     # Now go to the second page.
+    query_params = pagination_query_params
     login
     get(:test_index,
-        params: { num_per_page: 10, page: 2, q: pagination_query_param })
+        params: { num_per_page: 10, page: 2 }.merge(query_params))
     assert_template("names/index")
-
     name_links = css_select(".list-group.name-index a")
     assert_equal(10, name_links.length)
     expected = Name.order(:sort_name).limit(10).offset(10).to_a
     assert_equal(expected.map(&:id), ids_from_links(name_links))
-
     url = @controller.url_with_query(controller: "/names", action: :show,
                                      id: expected.first.id, only_path: true)
-    assert_equal(name_links.first[:href], url)
+    assert_not_nil(name_links.first.to_s.index(url))
 
+    # assert_select("a", text: "2", count: 0)
     assert_link_in_html("Previous", controller: "/names",
                                     action: :test_index, num_per_page: 10,
-                                    q: pagination_query_param, page: 1)
+                                    params: query_params, page: 1)
+    # assert_select("a", text: "Z", count: 0)
+    # assert_link_in_html("A", controller: "/names",
+    #                          action: :test_index, num_per_page: 10,
+    #                          params: query_params, letter: "A")
   end
 
   def test_pagination_letter
     # Now try a letter.
+    query_params = pagination_query_params
     l_names = Name.where(Name[:text_name].matches("L%")).
               order(:text_name, :author).to_a
     login
     get(:test_index, params: { num_per_page: l_names.size,
-                               letter: "L", q: pagination_query_param })
+                               letter: "L" }.merge(query_params))
     assert_template("names/index")
     assert_select("#content")
     name_links = css_select(".list-group.name-index a")
@@ -422,18 +432,24 @@ class NamesControllerIndexTest < FunctionalTestCase
 
     url = @controller.url_with_query(controller: "/names", action: :show,
                                      id: l_names.first.id, only_path: true)
-    assert_equal(name_links.first[:href], url)
+    assert_not_nil(name_links.first.to_s.index(url))
     assert_select("a", text: "1", count: 0)
+
+    # assert_select("a", text: "Z", count: 0)
+    # assert_link_in_html("A", controller: "/names",
+    #                          action: :test_index, params: query_params,
+    #                          num_per_page: l_names.size, letter: "A")
   end
 
   def test_pagination_letter_with_page
+    query_params = pagination_query_params
     l_names = Name.where(Name[:text_name].matches("L%")).
               order(:text_name, :author).to_a
     # Do it again, but make page size exactly one too small.
     l_names.pop
     login
     get(:test_index, params: { num_per_page: l_names.size,
-                               letter: "L", q: pagination_query_param })
+                               letter: "L" }.merge(query_params))
     assert_template("names/index")
     name_links = css_select(".list-group.name-index a")
 
@@ -441,39 +457,46 @@ class NamesControllerIndexTest < FunctionalTestCase
     assert_equal(Set.new(l_names.map(&:id)),
                  Set.new(ids_from_links(name_links)))
 
-    assert_link_in_html("Next", controller: "/names", action: :test_index,
-                                q: pagination_query_param,
+    # assert_select("a", text: "1", count: 0)
+    assert_link_in_html("Next", controller: "/names",
+                                action: :test_index, params: query_params,
                                 num_per_page: l_names.size,
                                 letter: "L", page: 2)
+    # assert_select("a", text: "3", count: 0)
   end
 
   def test_pagination_letter_with_page2
+    query_params = pagination_query_params
     l_names = Name.where(Name[:text_name].matches("L%")).
               order(:text_name, :author).to_a
     last_name = l_names.pop
     # Check second page.
     login
     get(:test_index, params: { num_per_page: l_names.size, letter: "L",
-                               page: 2, q: pagination_query_param })
+                               page: 2 }.merge(query_params))
     assert_template("names/index")
     name_links = css_select(".list-group.name-index a")
     assert_equal(1, name_links.length)
     assert_equal([last_name.id], ids_from_links(name_links))
-
-    assert_link_in_html("Previous", controller: "/names", action: :test_index,
-                                    q: pagination_query_param,
+    # assert_select("a", text: "2", count: 0)
+    assert_link_in_html("Previous", controller: "/names",
+                                    action: :test_index, params: query_params,
                                     num_per_page: l_names.size,
                                     letter: "L", page: 1)
+    # assert_select("a", text: "3", count: 0)
   end
 
   def test_pagination_with_anchors
+    query_params = pagination_query_params
     # Some cleverness is required to get pagination links to include anchors.
     login
-    get(:test_index, params: { num_per_page: 10, test_anchor: "blah",
-                               q: pagination_query_param })
+    get(:test_index, params: {
+      num_per_page: 10,
+      test_anchor: "blah"
+    }.merge(query_params))
     assert_link_in_html("Next", controller: "/names",
                                 action: :test_index, num_per_page: 10,
-                                q: pagination_query_param, page: 2,
+                                params: query_params, page: 2,
                                 test_anchor: "blah", anchor: "blah")
     # assert_link_in_html("A", controller: "/names",
     #                          action: :test_index, num_per_page: 10,
